requirements/base.txt
blinker
cffi==1.10.0
cookies==2.2.1
dnspython==1.15.0
future==0.16.0
hrt==0.1.0
# Here till they upload a Python3 compatible version of the library
ipaddr
Markdown==2.6.9
pexpect==4.6.0
psutil==5.3.1
psycopg2-binary==2.7.4
PTP==0.4.2
pycurl==7.43.0.2
pyOpenSSL==17.5.0
PyVirtualDisplay==0.2.1
PyYAML==3.13
requests==2.18.4
selenium==3.4.3
six==1.10.0
SQLAlchemy==1.1.13
sqlalchemy_mixins==1.1
tornado==5.0.2
